requirements.txt
bleak==0.22.3
pycryptodome==3.23.0
winrt-runtime==2.3.0
winrt-Windows.Devices.Bluetooth==2.3.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.3.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.3.0
winrt-Windows.Devices.Enumeration==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Storage.Streams==2.3.0
